backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.7
inflection==0.5.1
Markdown==3.5.2
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.5.0
tzdata==2023.4
urllib3==2.1.0
python-decouple==3.8
reportlab==4.2.0
PyPDF2==3.0.1
pytest==8.2.2
pytest-django==4.8.0
factory_boy==3.3.0
